Move PollyHelper to Internal folder/namespace
